Minor-Major Update to the Tile Engine.
New Classes:
Map
Avatar

Cleaned up large portion of the Code using new Naming Conventions to
increase readbility.

Added auto documentation to many of the current code files.

Added several TODOs for anyone who wants to help.
If you need to find them just Ctrl F for #TODO or TODO.

On the new classes:
Map.cs
Basic Object for keeping trace of the drawable Layers and the collision
layer.
Currently the player is stored in the map but I may remove it at a later
time.

Avatar.cs
The abstract type for a player.  Might just become the abstract type for
a dynamic movable object in the future.