added saving and loading to the mapeditor
just remember to put .xml as the extension as you file name when saving.
I might fix that later..

If you want these maps to be saved and synced to github, save them in a
different folder. Capstonet/xmlFiles is where our game reads them from
if you want to save them there